@@ -1,10 +1,9 @@
-# encoding: UTF-8
-
 require 'open-uri'
 require 'net/https'
 
 class PriceList::Parser
-  attr_accessor :url, :title, :price, :currency, :stock_state, :doc
+  attr_accessor :url, :price, :currency, :doc
+  attr_reader :title, :stock_state
 
   Dir.glob(File.join(File.dirname(__FILE__), 'parser', '*.rb')).each do |file|
     autoload File.basename(file, '.rb').camelize.to_sym, file
@@ -16,7 +15,36 @@ class PriceList::Parser
   end
 
   def parse
-    fail PriceList::Exception.new('extend me')
+    raise PriceList::Exception.new('extend me')
+  end
+
+  def uri
+    URI(url)
+  end
+
+  def html
+    pp url
+    URI(url).read(
+      'User-Agent' => 'Mozilla/5.0 (Windows NT 10.0; rv:131.0) Gecko/20100101 Firefox/131.0'
+    )
+  end
+
+  def doc
+    Nokogiri::HTML(html)
+  end
+
+  def title=(string)
+    @title = compact(string)
+  end
+
+  def stock_state=(string)
+    @stock_state = compact(string)
+  end
+
+  protected
+
+  def compact(string)
+    string&.strip&.gsub(/\n|\t/, ' ')&.squeeze(' ')
   end
 
   class << self
@@ -28,22 +56,23 @@ class PriceList::Parser
 
     def favicon_url
       checked_url = "#{responsible_base_url}/favicon.ico"
-      self.fetched_favicon_url ||= self.url_exists?(checked_url) ? checked_url : nil
+      self.fetched_favicon_url ||= url_exists?(checked_url) ? checked_url : nil
+      self.fetched_favicon_url ||= url_exists?(fetch_favicon) ? fetch_favicon : nil
       self.fetched_favicon_url
     end
 
     def uri
-      URI(self.responsible_base_url)
+      URI(responsible_base_url)
     end # .uri
 
     def responsible_regexp
-       %r{#{Regexp.escape(self.uri.host).gsub('www','.*')}}
+      /#{Regexp.escape(uri.host).gsub('www', '.*')}/
     end # #responsible_regexp
 
     def responsible?(url)
-      pp self.responsible_regexp unless PriceList.test?
+      pp responsible_regexp unless PriceList.test?
       pp url unless PriceList.test?
-      url =~ self.responsible_regexp
+      url =~ responsible_regexp
     end
 
     def responsible_class_name(url)
@@ -51,10 +80,22 @@ class PriceList::Parser
       constants.each do |constant|
         puts "Probe: #{constant}" unless PriceList.test?
         responsible = module_eval(constant.to_s).responsible?(url)
-        puts "\t#{responsible}"   unless PriceList.test?
+        puts "\t#{responsible}" unless PriceList.test?
         return "#{name}::#{constant}" if responsible
       end
-      fail PriceList::Exception.new('no matching parser found for given url')
+      raise PriceList::Exception.new('no matching parser found for given url')
+    end
+
+    def fetch_favicon
+      Nokogiri::HTML(
+        uri.read
+      ).css(
+        'link[rel="apple-touch-icon"]'
+      )&.first&.attr('href')
+    rescue Exception => e
+      puts e
+      puts e.backtrace * "\n"
+      nil
     end
 
     def url_exists?(_url)
@@ -70,9 +111,9 @@ class PriceList::Parser
 
       case res
       when Net::HTTPSuccess, Net::HTTPRedirection
-        return true
+        true
       else
-        return false
+        false
       end
     end
   end #  class << self
